Moved test component
